chore(sdk-core): move interfaces into folders
- interfaces & implementations will be grouped in same folder

Ticket: BG-48171
