+++ modules/sdk-core/src/bitgo/baseCoin/index.ts
@@ -1,0 +1,1 @@
+export * from './iBaseCoin';
