+++ modules/sdk-core/src/bitgo/enterprise/index.ts
@@ -1,0 +1,2 @@
+export * from './iEnterprise';
+export * from './iEnterprises';
