+++ modules/sdk-core/src/bitgo/index.ts
@@ -1,14 +1,14 @@
+export * from './baseCoin';
 export * from './bitgoBase';
+export * from './enterprise';
 export * from './environments';
-export * from './iBaseCoin';
-export * from './iEnterprise';
-export * from './iEnterprises';
-export * from './iKeychains';
-export * from './iMarkets';
-export * from './iPendingApproval';
-export * from './iPendingApprovals';
-export * from './iWallet';
-export * from './iWallets';
-export * from './iWebhooks';
+export * from './keychain';
+export * from './market';
+export * from './pendingApproval';
+export * from './recovery';
+export * from './trading';
 export * from './trading';
 export * from './types';
+export * from './utils';
+export * from './wallet';
+export * from './webhook';
