+++ modules/sdk-core/src/bitgo/keychain/index.ts
@@ -1,0 +1,1 @@
+export * from './iKeychains';
