+++ modules/sdk-core/src/bitgo/market/index.ts
@@ -1,0 +1,1 @@
+export * from './iMarkets';
