+++ modules/sdk-core/src/bitgo/pendingApproval/index.ts
@@ -1,0 +1,2 @@
+export * from './iPendingApproval';
+export * from './iPendingApprovals';
