+++ modules/sdk-core/src/bitgo/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './iTssUtils';
