+++ modules/sdk-core/src/bitgo/wallet/index.ts
@@ -1,0 +1,2 @@
+export * from './iWallet';
+export * from './iWallets';
