+++ modules/sdk-core/src/bitgo/webhook/index.ts
@@ -1,0 +1,1 @@
+export * from './iWebhooks';
